Resolve conflicts, remove duplicate code, and prepare for UX implementation
