+++ cursor_state.json
@@ -1,5 +1,5 @@
 {
   "cursor_session": {
-    "disconnected_at": "2025-08-15T18:04:42.170068"
+    "disconnected_at": "2025-08-15T23:02:59.821064"
   }
 }
