+++ memory-bank/DOCUMENTS/ux_first_actionable_20250116_phase0_postreview.md
@@ -1,0 +1,78 @@
+# Phase 0 Post-Review: SETUP & CONFLICT RESOLUTION
+**Task ID:** ux_first_actionable_20250116  
+**Phase:** 0  
+**Date:** 2025-01-16  
+**Status:** ✅ Complete
+
+## Conflicts Resolved
+
+### 1. ✅ Duplicate `/metrics` Endpoint
+- **Location:** `KYC VERIFICATION/src/api/app.py`
+- **Issue:** Two endpoints defined at lines 943 and 1464
+- **Resolution:** Removed duplicate endpoint at line 1464 (Prometheus-specific version)
+- **Result:** Single `/metrics` endpoint at line 943 remains
+
+### 2. ⚠️ SessionState vs EnhancedSessionState Conflict
+- **Location:** `KYC VERIFICATION/src/face/handlers.py`
+- **Issue:** Basic `SessionState` class conflicts with `EnhancedSessionState` from session_manager.py
+- **Actions Taken:**
+  - ✅ Removed `SessionState` class definition (lines 59-77)
+  - ✅ Added import for `EnhancedSessionState` from session_manager
+  - ✅ Updated `_sessions` dictionary type to use `EnhancedSessionState`
+  - ✅ Updated `get_or_create_session` function signature
+  - ✅ Updated session cleanup logic to use `created_at` instead of `last_update`
+
+## ✅ All Issues Resolved
+
+### SessionState Integration Complete
+Successfully integrated `EnhancedSessionState` with handlers.py by:
+
+1. **Added missing fields to EnhancedSessionState:**
+   - ✅ `stability_tracker` - For face stability tracking
+   - ✅ `pad_scores` (List) - For multiple PAD score storage
+   - ✅ `challenge_script` - For liveness challenge details
+   - ✅ `challenge_completed` - For challenge status
+   - ✅ `burst_frames` - For captured burst frames
+   - ✅ `lock_achieved_at` - For lock timestamp
+
+2. **Updated handlers.py integration:**
+   - ✅ Imported `EnhancedSessionState` from session_manager
+   - ✅ Updated `_sessions` dictionary type
+   - ✅ Modified `get_or_create_session` function
+   - ✅ Initialized `StabilityTracker` in new sessions
+   - ✅ Updated session cleanup logic
+
+## Validation Against IMPORTANT NOTE
+
+Per Phase 0 IMPORTANT NOTE: "These conflicts must be resolved first to prevent runtime errors. The duplicate endpoint will cause FastAPI to fail on startup."
+
+- ✅ **Duplicate endpoint resolved** - FastAPI will no longer fail on startup
+- ✅ **SessionState conflict resolved** - EnhancedSessionState extended with all required fields
+- ✅ **UX requirements documented** - All requirements A-H fully documented
+- ✅ **Backup created** - Implementation files backed up to `memory-bank/DOCUMENTS/kyc_implementation_backup_20250116.tar.gz`
+
+## Completed Actions
+
+1. **Conflict Resolution:**
+   - Removed duplicate `/metrics` endpoint at line 1464 in app.py
+   - Extended `EnhancedSessionState` with missing fields for full compatibility
+   - Updated all references to use `EnhancedSessionState`
+
+2. **Documentation:**
+   - Created comprehensive UX requirements documentation (A-H)
+   - Documented state flow and timing requirements
+   - Listed all Tagalog microcopy templates
+   - Defined acceptance criteria and parity checklist
+
+3. **Backup & Safety:**
+   - Created backup archive of modified files
+   - Preserved working implementation state
+   - Ready for Phase 1 implementation
+
+## Ready for Next Phase
+
+Phase 0 is now complete. The system is ready to proceed with:
+- **Phase 1**: State Machine Implementation (Requirement A)
+- All conflicts resolved
+- No runtime errors expected
+- Full UX requirements understood and documented
