+++ memory-bank/DOCUMENTS/ux_first_actionable_20250116_phase1_preanalysis.md
@@ -1,0 +1,98 @@
+# Phase 1 Pre-Analysis: STATE MACHINE IMPLEMENTATION
+**Task ID:** ux_first_actionable_20250116  
+**Phase:** 1  
+**Date:** 2025-01-16  
+**Status:** Ready to Execute
+
+## Phase 1 IMPORTANT NOTE (Restated)
+"State machine is the foundation for all UX flows. Must support both front and back document capture with proper transitions."
+
+## Phase Objectives
+Implement document capture state machine in EnhancedSessionState supporting:
+- SEARCHING→LOCKED→COUNTDOWN→CAPTURED→CONFIRM→FLIP-TO-BACK→BACK-SEARCHING→COMPLETE states
+- Add current_side tracking (front/back)
+- Add front_captured/back_captured flags  
+- Implement state transition methods
+- Ensure each state change emits proper telemetry events
+
+## Prerequisites ✅
+1. **EnhancedSessionState class exists** - Confirmed in session_manager.py
+2. **Handlers.py uses EnhancedSessionState** - Integration complete
+3. **Telemetry system available** - telemetry.py exists in face directory
+4. **UX requirements documented** - Requirements A-H documented
+
+## Risks & Mitigations
+
+### Risk 1: Breaking Existing Flows
+**Impact:** High  
+**Mitigation:** 
+- Implement state machine as extension, not replacement
+- Maintain backward compatibility with existing session fields
+- Add comprehensive unit tests before integration
+
+### Risk 2: State Transition Complexity
+**Impact:** Medium  
+**Mitigation:**
+- Create state transition matrix documentation
+- Validate transitions before allowing them
+- Log all state changes for debugging
+
+### Risk 3: Telemetry Event Overhead
+**Impact:** Low  
+**Mitigation:**
+- Use async event emission where possible
+- Batch telemetry events if needed
+- Monitor performance impact
+
+## Implementation Strategy
+
+### Step 1: Define State Enum
+- Create `CaptureState` enum with all states
+- Add to session_manager.py
+
+### Step 2: Extend EnhancedSessionState
+- Add `capture_state` field (default: SEARCHING)
+- Add `current_side` field (front/back)
+- Add `front_captured` and `back_captured` booleans
+- Add `state_history` list for tracking
+
+### Step 3: Implement State Transitions
+- Create `transition_to()` method with validation
+- Define allowed transitions matrix
+- Emit telemetry events on transitions
+
+### Step 4: Update Handlers
+- Modify lock_check to transition to LOCKED
+- Update burst_eval to handle CAPTURED state
+- Add flip-to-back logic
+
+### Step 5: Testing
+- Unit tests for all state transitions
+- Integration tests with handlers
+- Performance testing with telemetry
+
+## Success Criteria
+- ✅ All 8 states implemented
+- ✅ State transitions validated
+- ✅ Telemetry events emitted
+- ✅ Front/back tracking working
+- ✅ No breaking changes to existing flows
+- ✅ Tests passing with >90% coverage
+
+## Files to Modify
+1. `KYC VERIFICATION/src/face/session_manager.py` - Main state machine
+2. `KYC VERIFICATION/src/face/handlers.py` - Integration points
+3. `KYC VERIFICATION/src/face/telemetry.py` - Event emission
+
+## Rollback Plan
+If issues arise:
+1. Revert session_manager.py changes
+2. Restore handlers.py from backup
+3. Document issues for resolution
+4. Re-analyze approach before retry
+
+## Notes
+- State machine is foundational for all subsequent phases
+- Must be robust and well-tested
+- Consider using Python's enum.Enum for states
+- Document all state transitions clearly
