+++ memory-bank/current-session.md
@@ -1,4 +1,4 @@
-# 📝 Current Cursor Session — 2025-08-15 18:04:42 UTC
+# 📝 Current Cursor Session — 2025-08-15 23:02:59 UTC
 
 | Field | Value |
 |-------|-------|
@@ -7,4 +7,7 @@
 | current_task | — |
 | progress | — |
 | last_activity | — |
-| disconnected_at | 2025-08-15T18:04:42.170068 |
+| disconnected_at | 2025-08-15T23:02:59.821064 |
+
+## 🕒 Open Tasks (Todo Manager)
+- **Actionable plan for UX requirements implementation (A-H) followed by system integration, addressing conflicts and adding missing features.** (15 todos left)
