+++ memory-bank/cursor_state.json
@@ -1,8 +1,8 @@
 {
   "cursor_session": {
-    "disconnected_at": "2025-08-16T02:04:42.169451+08:00",
-    "current_task": "Actionable plan for backend-only face scan implementation compiled from organize.md (frozen source).",
-    "progress": 1.0,
-    "last_activity": "2025-08-16T02:04:42.157827+08:00"
+    "disconnected_at": "2025-08-16T07:02:59.820739+08:00",
+    "current_task": "Actionable plan for UX requirements implementation (A-H) followed by system integration, addressing conflicts and adding missing features.",
+    "progress": 0.0625,
+    "last_activity": "2025-08-16T07:02:59.798002+08:00"
   }
 }
