+++ memory-bank/task_interruption_state.json
@@ -1,5 +1,5 @@
 {
-  "current_task": "face_scan_actionable_20250116",
+  "current_task": "ux_first_actionable_20250116",
   "interrupted_tasks": [],
-  "last_updated": "2025-08-16T02:04:42.169884+08:00"
+  "last_updated": "2025-08-16T07:02:59.820919+08:00"
 }
