+++ memory-bank/task_state.json
@@ -1,8 +1,8 @@
 {
-  "current_task_id": "face_scan_actionable_20250116",
-  "current_task_description": "Actionable plan for backend-only face scan implementation compiled from organize.md (frozen source).",
-  "status": "completed",
-  "last_updated": "2025-08-16T02:04:42.157827+08:00",
-  "todos_count": 13,
-  "completed_todos": 13
+  "current_task_id": "ux_first_actionable_20250116",
+  "current_task_description": "Actionable plan for UX requirements implementation (A-H) followed by system integration, addressing conflicts and adding missing features.",
+  "status": "in_progress",
+  "last_updated": "2025-08-16T07:02:59.798002+08:00",
+  "todos_count": 16,
+  "completed_todos": 1
 }
